+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KGPKW7CAZDCFW17XX9SPCS05"
+        "01KGPRQS2FK7YDWE30K4X2TCGF"
     )
     r"""The unique identifier of the data source"""