+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasetop.py
@@ -81,7 +81,7 @@ class RetrieveDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-02-05T09:59:17.965Z")
+    updated: Optional[datetime] = parse_datetime("2026-02-05T11:24:14.219Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")