+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KGPKW7C6K10273RQXW4X6H0Y"
+        "01KGPRQS2DYA810CHQFJY3358E"
     )
     r"""The unique identifier of the data source"""