+++ packages/orq-rc/src/orq_ai_sdk/models/updateconversationop.py
@@ -231,7 +231,7 @@ class UpdateConversationResponseBody(BaseModel):
     r"""Unix timestamp in milliseconds when the conversation was last modified."""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kgpkw6sfcvzc5q25s5hv2hhx"
+        "conv_01kgprqrdrkpdgd41ztsw1kztv"
     )
     r"""Unique conversation identifier with `conv_` prefix."""