+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KGPKW7CEWGTFEXNBKB7VK6NT"
+        "01KGPRQS2J32REQKKBK7HPD31D"
     )
     r"""The unique identifier of the data source"""